fix: dollar sign; up: parser.c has been splitted
